Account for new gpuArray support in selectStrongestBboxes
Fixes #20.

selectStrongestBbox got gpuArray support in MATLAB R2020b. Previously the gather was required if bboxes was gpuArray and needed to be on the cpu (strangely the scores didn't matter).

In R2020b they need to be the same either both on the gpu or not, so we
do a version check and adjust behaviour accordingly